requirements.txt
langchain-community
fastapi
uvicorn
langchain
langchain-openai
langsmith
python-dotenv
requests
faiss-cpu
pypdf
certifi

